pyproject.toml
[project]
name = "autosweep"
version = "0.1.0"
description = "Automated daily sweeps 4tw"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fake-useragent>=2.2.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "selenium>=4.31.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]
